requirements.txt
numpy<2
faiss-cpu==1.8.0
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
sentence-transformers==3.0.1
pypdf==4.3.1
langchain-text-splitters==0.2.2
requests==2.32.3
